Handle SoundCloud URLs passed to “Tune In URL”
A number of plugins allow you to give them a webpage URL, and they’ll play the music associated with that webpage, e.g., the Band’s Campout plugin has a “Bandcamp URL” field, the YouTube plugin has a “YouTube URL or Video id” field, etc. However, it would be more convenient if you could paste the URL into the “Tune In URL” field and have LMS work out for you which plugin to use.

This commit provides that for SoundCloud tracks.

Support for this in LMS landed on the public/8.0 branch today, but this change should safely do nothing on older versions.